+++ src/index.ts
@@ -7,55 +7,55 @@ import {
 	User
 } from "discord.js";
 
-type PaginatorData = {
-
+export type PaginatorData = {
 	/**
 	 * The text to be displayed when using embed pages.
 	 */
 	content: string | null;
-
 	/**
 	 * Embed template to use across pages.
 	 * All properties on this object will be shown on every page.
 	 */
 	embed: MessageEmbed | null;
-
 	/**
 	 * The emojis to use for skipping the pages.
 	 * Specify the id for custom guild emojis.
 	 */
 	emojis: Partial<PaginatorEmojis> | null;
-
 	/**
 	 * The id of the user to listen to.
 	 */
 	userID: string | null;
-
 	/**
 	 * The time in milliseconds of how long to keep the paginator running.
 	 * @default 120000
 	 */
 	timeout: number;
-
 	/**
-	 * The pages to use in the paginator. Must be 2 or more pages.
-	 */
-	pages: Array<string | MessageEmbed>;
-
-	/**
-	 * If set to `true`, the paginator will include the stop emoji.
+	 * Whether or not to delete the paginator when timed out.
 	 * @default false
 	 */
-	stoppable: boolean;
-
+	deleteOnTimeout: boolean;
+	/**
+	 * The pages to use in the paginator. Must contain at least one page.
+	 */
+	pages: Array<string | MessageEmbed>;
 	/**
 	 * If set to `true`, the pages will loop on 'back' and 'next.'
 	 * @default true
 	 */
 	circular: boolean;
+	/**
+	 * Options to use for the 'info' emoji.
+	 */
+	infoOptions: PaginatorInfoOptions;
+	/**
+	 * Options to use for the 'jump' emoji.
+	 */
+	jumpOptions: PaginatorJumpOptions;
 };
 
-type PaginatorEmojis = {
+export type PaginatorEmojis = {
 
 	/** @default "⏮" */
 	front: string | null;
@@ -82,11 +82,38 @@ type PaginatorEmojis = {
 	info: string | null;
 };
 
-interface PaginatorEvents {
-	end: [string];
-	error: [Error];
-}
+export type PaginatorInfoOptions = {
+	/**
+	 * @default "This is a paginator. React to change page."
+	 */
+	text: string;
+	/**
+	 * @default 5000
+	 */
+	timeout: number;
+};
 
+export type PaginatorJumpOptions = {
+	/**
+	 * @default "Enter page number to jump to."
+	 */
+	prompt: string;
+	/**
+	 * Whether or not to show prompt.
+	 * @default true
+	 */
+	showPrompt: boolean;
+	/**
+	 * @default true
+	 */
+	deleteResponse: boolean;
+	/**
+	 * @default 10000
+	 */
+	timeout: number;
+};
+
+/** @internal */
 interface MessageEmbed {
 	title?: string | null;
 	description?: string | null;
@@ -115,6 +142,7 @@ interface MessageEmbed {
 	}[];
 }
 
+/** @internal */
 const PaginatorEmojisDefault: Readonly<PaginatorEmojis> = {
 	front: "⏮",
 	rear: "⏭",
@@ -126,12 +154,27 @@ const PaginatorEmojisDefault: Readonly<PaginatorEmojis> = {
 	info: "ℹ️"
 };
 
+/** @internal */
+const PaginatorInfoOptionsDefault: Readonly<PaginatorInfoOptions> = {
+	text: "This is a paginator. React to change page.",
+	timeout: 5000
+};
+
+/** @internal */
+const PaginatorJumpOptionsDefault: Readonly<PaginatorJumpOptions> = {
+	prompt: "Enter page number to jump to.",
+	showPrompt: true,
+	deleteResponse: true,
+	timeout: 10000
+};
+
+/** @internal */
 function delay(timeout: number): Promise<void> {
 	return new Promise(resolve => setTimeout(resolve, timeout));
 }
 
 export class Paginator extends EventEmitter {
-	public destroyed: boolean;
+	public destroyed = false;
 
 	private static readonly _footerFormat = "Page {0}/{1}";
 
@@ -146,9 +189,11 @@ export class Paginator extends EventEmitter {
 	private _embed: MessageEmbed | null;
 	private _emojis: PaginatorEmojis;
 	private _pages: Array<string | MessageEmbed>;
-	private _stoppable: boolean;
 	private _timeout: number;
+	private _deleteOnTimeout: boolean;
 	private _userID: string | null;
+	private _infoOptions: PaginatorInfoOptions;
+	private _jumpOptions: PaginatorJumpOptions;
 	private _pageCount: number;
 	private _lastPageIndex: number;
 
@@ -161,21 +206,23 @@ export class Paginator extends EventEmitter {
 			embed = null,
 			emojis = null,
 			pages = [],
-			stoppable = false,
 			timeout = 120000,
-			userID = null
+			deleteOnTimeout = false,
+			userID = null,
+			infoOptions = null,
+			jumpOptions = null
 		} = data;
-
-		this.destroyed = false;
 
 		this._circular = circular;
 		this._content = content;
 		this._embed = embed;
 		this._emojis = { ...PaginatorEmojisDefault, ...emojis };
 		this._pages = pages;
-		this._stoppable = stoppable;
 		this._timeout = timeout;
+		this._deleteOnTimeout = deleteOnTimeout;
 		this._userID = userID;
+		this._infoOptions = { ...PaginatorInfoOptionsDefault, ...infoOptions };
+		this._jumpOptions = { ...PaginatorJumpOptionsDefault, ...jumpOptions };
 		this._pageCount = this._pages.length;
 		this._lastPageIndex = this._pageCount - 1;
 	}
@@ -199,13 +246,23 @@ export class Paginator extends EventEmitter {
 		return this;
 	}
 
+	public setInfoOptions(infoOptions: Partial<PaginatorInfoOptions>): this {
+		this._infoOptions = { ...PaginatorInfoOptionsDefault, ...infoOptions };
+		return this;
+	}
+
+	public setJumpOptions(jumpOptions: Partial<PaginatorJumpOptions>): this {
+		this._jumpOptions = { ...PaginatorJumpOptionsDefault, ...jumpOptions };
+		return this;
+	}
+
 	public setTimeout(timeout: number): this {
 		this._timeout = timeout;
 		return this;
 	}
 
-	public setStoppable(stoppable: boolean): this {
-		this._stoppable = stoppable;
+	public setDeleteOnTimeout(deleteOnTimeout: boolean): this {
+		this._deleteOnTimeout = deleteOnTimeout;
 		return this;
 	}
 
@@ -248,12 +305,11 @@ export class Paginator extends EventEmitter {
 			this._emojis.back,
 			this._emojis.next,
 			this._emojis.rear,
+			this._emojis.stop,
 			this._emojis.jump,
 			this._emojis.trash,
 			this._emojis.info
 		].filter(x => typeof x === "string");
-
-		if (this._stoppable && this._emojis.stop) emojis.splice(4, 0, this._emojis.stop);
 
 		for (const emoji of emojis) {
 			await this._message!.react(emoji!);
@@ -272,23 +328,6 @@ export class Paginator extends EventEmitter {
 		this.destroyed = true;
 		this._collector!.stop();
 		this.emit("end", "Paginator was destroyed.");
-	}
-
-	public on<K extends keyof PaginatorEvents>(event: K, listener: (...args: PaginatorEvents[K]) => void): this;
-	public on(event: string | symbol, listener: (...args: any[]) => void): this {
-		super.on(event, listener);
-		return this;
-	}
-
-	public once<K extends keyof PaginatorEvents>(event: K, listener: (...args: PaginatorEvents[K]) => void): this;
-	public once(event: string | symbol, listener: (...args: any[]) => void): this {
-		super.once(event, listener);
-		return this;
-	}
-
-	public emit<K extends keyof PaginatorEvents>(event: K, ...args: PaginatorEvents[K]): boolean;
-	public emit(event: string | symbol, ...args: any[]): boolean {
-		return super.emit(event, ...args);
 	}
 
 	private async _setMessage(channel: TextBasedChannelFields): Promise<void> {
@@ -401,6 +440,20 @@ export class Paginator extends EventEmitter {
 			return false;
 		}
 
+		if (typeof this._infoOptions.text !== "string") {
+			this.emit("error", new TypeError(
+				`infoOptions.text must be of type 'string.' Received type '${typeof this._infoOptions.text}.'`
+			));
+			return false;
+		}
+
+		if (typeof this._jumpOptions.prompt !== "string") {
+			this.emit("error", new TypeError(
+				`jumpOptions.prompt must be of type 'string.' Received type '${typeof this._jumpOptions.prompt}.'`
+			));
+			return false;
+		}
+
 		return true;
 	}
 
@@ -470,18 +523,22 @@ export class Paginator extends EventEmitter {
 			}
 
 			case this._emojis.info: {
-				const msg = await this._message!.channel.send("This is a paginator. React to change page.");
-				await msg.delete({ timeout: 5000 });
+				const msg = await this._message!.channel.send(this._infoOptions.text);
+				await msg.delete({ timeout: this._infoOptions.timeout });
 				return;
 			}
 
 			case this._emojis.jump: {
-				const msg = await this._message!.channel.send("Enter page number to jump to.");
-				const filter = (x: Message): boolean => x.author.id === this._userID;
-				const collected = await msg.channel.awaitMessages(filter, { max: 1, time: 10000 });
+				const promptMessage = this._jumpOptions.showPrompt
+					? await this._message!.channel.send(this._jumpOptions.prompt)
+					: null;
 
-				await msg.delete();
+				const collected = await this._message!.channel.awaitMessages(
+					x => x.author.id === this._userID,
+					{ max: 1, time: this._jumpOptions.timeout }
+				);
 
+				if (promptMessage !== null) await promptMessage.delete();
 				if (collected.size === 0) return;
 
 				const collectedMessage = collected.first()!;
@@ -489,7 +546,7 @@ export class Paginator extends EventEmitter {
 
 				if (Number.isNaN(pageNumber)) return;
 				if (pageNumber <= 0 || pageNumber > this._pageCount) return;
-				if (hasManageMessages) await collectedMessage.delete();
+				if (this._jumpOptions.deleteResponse && hasManageMessages) await collectedMessage.delete();
 
 				this._currentPageIndex = pageNumber - 1;
 				break;
@@ -502,6 +559,23 @@ export class Paginator extends EventEmitter {
 	private _onEnd(): void {
 		this._running = false;
 
-		if (!this.destroyed) this.emit("end", "Paginator timed out.");
+		if (!this.destroyed) {
+			if (this._deleteOnTimeout) this._message!.delete();
+			this.emit("end", "Paginator timed out.");
+		}
 	}
+}
+
+export interface Paginator {
+	on(event: "end", listener: (reason: string) => void): this;
+	on(event: "error", listener: (error: Error | TypeError) => void): this;
+
+	once(event: "end", listener: (reason: string) => void): this;
+	once(event: "error", listener: (error: Error | TypeError) => void): this;
+
+	off(event: "end", listener: (reason: string) => void): this;
+	off(event: "error", listener: (error: Error | TypeError) => void): this;
+
+	emit(event: "end", reason: string): boolean;
+	emit(event: "error", error: Error | TypeError): boolean;
 }