+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 
+## 2.0.0
+### Added
+- Add option to delete on timeout
+- Added info options
+- Added jump options
+
+### Removed
+- Removed redundant stoppable option
+
 ## 1.1.0
 ### Added
-- Support trash, jump and info
+- Support trash, jump and info emojis
 - Support content when using embeds
 - Now checks channel permission for MANAGE_MESSAGES
 - Now works in DM
@@ -11,4 +20,4 @@
 - Paginator no longer emits error when 1 page is given
 
 ## 1.0.0
-- Initial release
+- Initial commit